+++ src/main.tsx
@@ -20,10 +20,12 @@ import "./styles/theme.css";
 import "./lsShimInstall";
 import "./styles/app.css";
 import { installEditableEmojiPlatform } from "./editableEmoji";
+import { applyContentWidths } from "./contentWidth";
 
 installEditableEmojiPlatform();
 applyTheme();
 applyAccent();
+applyContentWidths();
 const communityExtensionsReady = startCommunityExtensions()
   .then(({ pluginInitialization }) => {
     void pluginInitialization.catch((error) =>